Update README.md
Trinity College Library is a website that hopes to inform users about historical facts about the old college library, The Book of Kells, and The Long Room. The goal of the website is to provide a reliable source of information about the college library, The Book of Kells, and Room Long.